+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- Added `_into` Datetime kernel variants for efficient non-allocations.
 - Added `_into` Bitmask kernel variants for efficient non-allocations.
 - Added UnsafeMut utility Buffer for handling cross-thread bitpacked writes.